+++ assets/icon/index.ts
@@ -22,3 +22,4 @@ export { default as AddCategoryCloseIcon } from './addCategoryClose.png';
 export { default as CategoryEditIcon } from './categoryEdit.png';
 export { default as CategoryDeleteIcon } from './categoryDelete.png';
 export { default as OpenMySizeIcon } from './openMysize.png';
+export { default as CategoryDetailFirstClosetIcon } from './blankCategoryDetail.png';
